Refactor generate function to a separate file
