+++ src/saml.ts
@@ -42,7 +42,12 @@ import {
 } from "./xml";
 import { certToPEM, generateUniqueId, keyToPEM } from "./crypto";
 import { dateStringToTimestamp, generateInstant } from "./datetime";
-import { generateServiceProviderMetadata } from "./saml/metadata";
+import {
+  generateAuthorizeRequestAsync,
+  generateServiceProviderMetadata,
+  _generateLogoutRequest,
+  _generateLogoutResponse,
+} from "./saml/generate";
 
 const inflateRawAsync = util.promisify(zlib.inflateRaw);
 const deflateRawAsync = util.promisify(zlib.deflateRaw);
@@ -188,7 +193,7 @@ class SAML {
     return options;
   }
 
-  private getCallbackUrl(host?: string | undefined): string {
+  protected getCallbackUrl(host?: string | undefined): string {
     // Post-auth destination
     if (this.options.callbackUrl) {
       return this.options.callbackUrl;
@@ -207,7 +212,7 @@ class SAML {
     }
   }
 
-  private signRequest(samlMessage: querystring.ParsedUrlQueryInput): void {
+  protected signRequest(samlMessage: querystring.ParsedUrlQueryInput): void {
     assertRequired(this.options.privateKey, "privateKey is required");
 
     const samlMessageToSign: querystring.ParsedUrlQueryInput = {};
@@ -229,250 +234,11 @@ class SAML {
     samlMessage.Signature = signer.sign(keyToPEM(this.options.privateKey), "base64");
   }
 
-  private async generateAuthorizeRequestAsync(
-    isPassive: boolean,
-    isHttpPostBinding: boolean,
-    host: string | undefined
-  ): Promise<string> {
-    assertRequired(this.options.entryPoint, "entryPoint is required");
+  protected generateAuthorizeRequestAsync = generateAuthorizeRequestAsync;
 
-    const id = this.options.generateUniqueId();
-    const instant = generateInstant();
+  _generateLogoutRequest = _generateLogoutRequest;
 
-    if (this.mustValidateInResponseTo(true)) {
-      await this.cacheProvider.saveAsync(id, instant);
-    }
-    const request: AuthorizeRequestXML = {
-      "samlp:AuthnRequest": {
-        "@xmlns:samlp": "urn:oasis:names:tc:SAML:2.0:protocol",
-        "@ID": id,
-        "@Version": "2.0",
-        "@IssueInstant": instant,
-        "@ProtocolBinding": "urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST",
-        "@Destination": this.options.entryPoint,
-        "saml:Issuer": {
-          "@xmlns:saml": "urn:oasis:names:tc:SAML:2.0:assertion",
-          "#text": this.options.issuer,
-        },
-      },
-    };
-
-    if (isPassive) request["samlp:AuthnRequest"]["@IsPassive"] = true;
-
-    if (this.options.forceAuthn === true) {
-      request["samlp:AuthnRequest"]["@ForceAuthn"] = true;
-    }
-
-    if (!this.options.disableRequestAcsUrl) {
-      request["samlp:AuthnRequest"]["@AssertionConsumerServiceURL"] = this.getCallbackUrl(host);
-    }
-
-    const samlAuthnRequestExtensions = this.options.samlAuthnRequestExtensions;
-    if (samlAuthnRequestExtensions != null) {
-      if (typeof samlAuthnRequestExtensions != "object") {
-        throw new TypeError("samlAuthnRequestExtensions should be Object");
-      }
-      request["samlp:AuthnRequest"]["samlp:Extensions"] = {
-        "@xmlns:samlp": "urn:oasis:names:tc:SAML:2.0:protocol",
-        ...samlAuthnRequestExtensions,
-      };
-    }
-
-    const nameIDPolicy: XMLInput = {
-      "@xmlns:samlp": "urn:oasis:names:tc:SAML:2.0:protocol",
-      "@AllowCreate": this.options.allowCreate,
-    };
-
-    if (this.options.identifierFormat != null) {
-      nameIDPolicy["@Format"] = this.options.identifierFormat;
-    }
-
-    if (this.options.spNameQualifier != null) {
-      nameIDPolicy["@SPNameQualifier"] = this.options.spNameQualifier;
-    }
-
-    request["samlp:AuthnRequest"]["samlp:NameIDPolicy"] = nameIDPolicy;
-
-    if (!this.options.disableRequestedAuthnContext) {
-      const authnContextClassRefs: XMLInput[] = [];
-      (this.options.authnContext as string[]).forEach(function (value) {
-        authnContextClassRefs.push({
-          "@xmlns:saml": "urn:oasis:names:tc:SAML:2.0:assertion",
-          "#text": value,
-        });
-      });
-
-      request["samlp:AuthnRequest"]["samlp:RequestedAuthnContext"] = {
-        "@xmlns:samlp": "urn:oasis:names:tc:SAML:2.0:protocol",
-        "@Comparison": this.options.racComparison,
-        "saml:AuthnContextClassRef": authnContextClassRefs,
-      };
-    }
-
-    if (this.options.attributeConsumingServiceIndex != null) {
-      request["samlp:AuthnRequest"]["@AttributeConsumingServiceIndex"] =
-        this.options.attributeConsumingServiceIndex;
-    }
-
-    if (this.options.providerName != null) {
-      request["samlp:AuthnRequest"]["@ProviderName"] = this.options.providerName;
-    }
-
-    if (this.options.scoping != null) {
-      const scoping: XMLInput = {
-        "@xmlns:samlp": "urn:oasis:names:tc:SAML:2.0:protocol",
-      };
-
-      if (typeof this.options.scoping.proxyCount === "number") {
-        scoping["@ProxyCount"] = this.options.scoping.proxyCount;
-      }
-
-      if (this.options.scoping.idpList) {
-        scoping["samlp:IDPList"] = this.options.scoping.idpList.map(
-          (idpListItem: SamlIDPListConfig) => {
-            const formattedIdpListItem: XMLInput = {
-              "@xmlns:samlp": "urn:oasis:names:tc:SAML:2.0:protocol",
-            };
-
-            if (idpListItem.entries) {
-              formattedIdpListItem["samlp:IDPEntry"] = idpListItem.entries.map(
-                (entry: SamlIDPEntryConfig) => {
-                  const formattedEntry: XMLInput = {
-                    "@xmlns:samlp": "urn:oasis:names:tc:SAML:2.0:protocol",
-                  };
-
-                  formattedEntry["@ProviderID"] = entry.providerId;
-
-                  if (entry.name) {
-                    formattedEntry["@Name"] = entry.name;
-                  }
-
-                  if (entry.loc) {
-                    formattedEntry["@Loc"] = entry.loc;
-                  }
-
-                  return formattedEntry;
-                }
-              );
-            }
-
-            if (idpListItem.getComplete) {
-              formattedIdpListItem["samlp:GetComplete"] = idpListItem.getComplete;
-            }
-
-            return formattedIdpListItem;
-          }
-        );
-      }
-
-      if (this.options.scoping.requesterId) {
-        scoping["samlp:RequesterID"] = this.options.scoping.requesterId;
-      }
-
-      request["samlp:AuthnRequest"]["samlp:Scoping"] = scoping;
-    }
-
-    let stringRequest = buildXmlBuilderObject(request, false);
-    // TODO: maybe we should always sign here
-    if (isHttpPostBinding && isValidSamlSigningOptions(this.options)) {
-      stringRequest = signAuthnRequestPost(stringRequest, this.options);
-    }
-    return stringRequest;
-  }
-
-  async _generateLogoutRequest(user: Profile): Promise<string> {
-    const id = this.options.generateUniqueId();
-    const instant = generateInstant();
-
-    const request = {
-      "samlp:LogoutRequest": {
-        "@xmlns:samlp": "urn:oasis:names:tc:SAML:2.0:protocol",
-        "@xmlns:saml": "urn:oasis:names:tc:SAML:2.0:assertion",
-        "@ID": id,
-        "@Version": "2.0",
-        "@IssueInstant": instant,
-        "@Destination": this.options.logoutUrl,
-        "saml:Issuer": {
-          "@xmlns:saml": "urn:oasis:names:tc:SAML:2.0:assertion",
-          "#text": this.options.issuer,
-        },
-        "samlp:Extensions": {},
-        "saml:NameID": {
-          "@Format": user.nameIDFormat,
-          "#text": user.nameID,
-        },
-      },
-    } as LogoutRequestXML;
-
-    const samlLogoutRequestExtensions = this.options.samlLogoutRequestExtensions;
-    if (samlLogoutRequestExtensions != null) {
-      if (typeof samlLogoutRequestExtensions != "object") {
-        throw new TypeError("samlLogoutRequestExtensions should be Object");
-      }
-      request["samlp:LogoutRequest"]["samlp:Extensions"] = {
-        "@xmlns:samlp": "urn:oasis:names:tc:SAML:2.0:protocol",
-        ...samlLogoutRequestExtensions,
-      };
-    } else {
-      delete request["samlp:LogoutRequest"]["samlp:Extensions"];
-    }
-
-    if (user.nameQualifier != null) {
-      request["samlp:LogoutRequest"]["saml:NameID"]["@NameQualifier"] = user.nameQualifier;
-    }
-
-    if (user.spNameQualifier != null) {
-      request["samlp:LogoutRequest"]["saml:NameID"]["@SPNameQualifier"] = user.spNameQualifier;
-    }
-
-    if (user.sessionIndex) {
-      request["samlp:LogoutRequest"]["saml2p:SessionIndex"] = {
-        "@xmlns:saml2p": "urn:oasis:names:tc:SAML:2.0:protocol",
-        "#text": user.sessionIndex,
-      };
-    }
-
-    await this.cacheProvider.saveAsync(id, instant);
-    return buildXmlBuilderObject(request, false);
-  }
-
-  _generateLogoutResponse(logoutRequest: Profile, success: boolean): string {
-    const id = this.options.generateUniqueId();
-    const instant = generateInstant();
-
-    const successStatus = {
-      "samlp:StatusCode": {
-        "@Value": "urn:oasis:names:tc:SAML:2.0:status:Success",
-      },
-    };
-
-    const failStatus = {
-      "samlp:StatusCode": {
-        "@Value": "urn:oasis:names:tc:SAML:2.0:status:Requester",
-        "samlp:StatusCode": {
-          "@Value": "urn:oasis:names:tc:SAML:2.0:status:UnknownPrincipal",
-        },
-      },
-    };
-
-    const request = {
-      "samlp:LogoutResponse": {
-        "@xmlns:samlp": "urn:oasis:names:tc:SAML:2.0:protocol",
-        "@xmlns:saml": "urn:oasis:names:tc:SAML:2.0:assertion",
-        "@ID": id,
-        "@Version": "2.0",
-        "@IssueInstant": instant,
-        "@Destination": this.options.logoutUrl,
-        "@InResponseTo": logoutRequest.ID,
-        "saml:Issuer": {
-          "#text": this.options.issuer,
-        },
-        "samlp:Status": success ? successStatus : failStatus,
-      },
-    };
-
-    return buildXmlBuilderObject(request, false);
-  }
+  _generateLogoutResponse = _generateLogoutResponse;
 
   async _requestToUrlAsync(
     request: string | null | undefined,
@@ -670,7 +436,7 @@ class SAML {
     )(callback);
   }
 
-  private async getLogoutResponseUrlAsync(
+  protected async getLogoutResponseUrlAsync(
     samlLogoutRequest: Profile,
     RelayState: string,
     options: AuthenticateOptions & AuthorizeOptions,
@@ -687,7 +453,7 @@ class SAML {
     );
   }
 
-  private async certsToCheck(): Promise<string[]> {
+  protected async certsToCheck(): Promise<string[]> {
     let checkedCerts: string[];
 
     if (typeof this.options.cert === "function") {
@@ -870,7 +636,7 @@ class SAML {
     }
   }
 
-  private async validateInResponseTo(inResponseTo: string | null): Promise<void> {
+  protected async validateInResponseTo(inResponseTo: string | null): Promise<void> {
     if (this.mustValidateInResponseTo(Boolean(inResponseTo))) {
       if (inResponseTo) {
         const result = await this.cacheProvider.getAsync(inResponseTo);
@@ -900,7 +666,7 @@ class SAML {
     return await processValidlySignedSamlLogoutAsync(doc, dom, this.options.decryptionPvk ?? null);
   }
 
-  private async hasValidSignatureForRedirect(
+  protected async hasValidSignatureForRedirect(
     container: ParsedQs,
     originalQuery: string
   ): Promise<boolean | void> {
@@ -938,7 +704,7 @@ class SAML {
     }
   }
 
-  private validateSignatureForRedirect(
+  protected validateSignatureForRedirect(
     urlString: crypto.BinaryLike,
     signature: string,
     alg: string,
@@ -965,7 +731,7 @@ class SAML {
     return verifier.verify(certToPEM(cert), signature, "base64");
   }
 
-  private verifyLogoutRequest(doc: XMLOutput): void {
+  protected verifyLogoutRequest(doc: XMLOutput): void {
     this.verifyIssuer(doc.LogoutRequest);
     const nowMs = new Date().getTime();
     const conditions = doc.LogoutRequest.$;
@@ -979,7 +745,7 @@ class SAML {
     }
   }
 
-  private async verifyLogoutResponse(doc: XMLOutput): Promise<void> {
+  protected async verifyLogoutResponse(doc: XMLOutput): Promise<void> {
     const statusCode = doc.LogoutResponse.Status[0].StatusCode[0].$.Value;
     if (statusCode !== "urn:oasis:names:tc:SAML:2.0:status:Success")
       throw new Error("Bad status code: " + statusCode);
@@ -993,7 +759,7 @@ class SAML {
     return;
   }
 
-  private verifyIssuer(samlMessage: XMLOutput): void {
+  protected verifyIssuer(samlMessage: XMLOutput): void {
     if (this.options.idpIssuer != null) {
       const issuer = samlMessage.Issuer;
       if (issuer) {
@@ -1007,7 +773,7 @@ class SAML {
     }
   }
 
-  private async processValidlySignedAssertionAsync(
+  protected async processValidlySignedAssertionAsync(
     xml: string,
     samlResponseXml: string,
     inResponseTo: string | null
@@ -1208,7 +974,7 @@ class SAML {
     return { profile, loggedOut: false };
   }
 
-  private checkTimestampsValidityError(
+  protected checkTimestampsValidityError(
     nowMs: number,
     notBefore: string,
     notOnOrAfter: string,
@@ -1234,7 +1000,7 @@ class SAML {
     return null;
   }
 
-  private checkAudienceValidityError(
+  protected checkAudienceValidityError(
     expectedAudience: string,
     audienceRestrictions: AudienceRestrictionXML[]
   ): Error | null {
@@ -1273,19 +1039,7 @@ class SAML {
     return await processValidlySignedPostRequestAsync(doc, dom, this.options.decryptionPvk ?? null);
   }
 
-  generateServiceProviderMetadata(
-    decryptionCert: string | null,
-    signingCerts?: string | string[] | null
-  ): string {
-    const callbackUrl = this.getCallbackUrl(); // TODO it would probably be useful to have a host parameter here
-
-    return generateServiceProviderMetadata({
-      ...this.options,
-      callbackUrl,
-      decryptionCert,
-      signingCerts,
-    });
-  }
+  generateServiceProviderMetadata = generateServiceProviderMetadata;
 
   /**
    * Process max age assertion and use it if it is more restrictive than the NotOnOrAfter age
@@ -1296,7 +1050,7 @@ class SAML {
    * @param issueInstant Time when response was issued.
    * @returns {*} The expiration time to be used, in Ms.
    */
-  private processMaxAgeAssertionTime(
+  protected processMaxAgeAssertionTime(
     maxAssertionAgeMs: number,
     notOnOrAfter: string,
     issueInstant: string
@@ -1312,7 +1066,7 @@ class SAML {
     return maxAssertionTimeMs < notOnOrAfterMs ? maxAssertionTimeMs : notOnOrAfterMs;
   }
 
-  private mustValidateInResponseTo(hasInResponseTo: boolean): boolean {
+  protected mustValidateInResponseTo(hasInResponseTo: boolean): boolean {
     return (
       this.options.validateInResponseTo === ValidateInResponseTo.always ||
       (this.options.validateInResponseTo === ValidateInResponseTo.ifPresent && hasInResponseTo)
